Capstone Project HR Analytics
How salary is varying with respect to other considerable
factors.

Visualize the insight of the data set by doing EDA.

Build the model and thereby apply it to test data to measure
the performance and robustness of the model.

Machine learning techniques like Decision tree and random
forest regression to check the accuracy of the model.

Based on the model and its accuracy, hr department
is recommended to make structural changes to get company
into more profit.